Move config to separate tab
